break up engine file, add hashmap classes
